requirements.txt
blinker==1.8.2
cachelib==0.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
circuitbreaker==2.0.0
click==8.1.7
cryptography==43.0.1
Deprecated==1.2.14
Faker==30.0.0
Flask==3.0.3
Flask-Caching==2.1.0
Flask-Limiter==3.8.0
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
limits==3.13.0
linecache2==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
mdurl==0.1.2
mysql-connector-python==9.0.0
ordered-set==4.1.0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
save==0.1
six==1.16.0
SQLAlchemy==2.0.35
traceback2==1.4.0
typing_extensions==4.12.2
unittest2==1.1.0
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.16.0
